Create README.md
A university computer science department has a teaching assistant (TA) who helps
undergraduate students with their programming assignments during regular office hours.
The TA’s office is rather small and has room for only one desk with a chair and
computer. There are three chairs in the hallway outside the office where students can sit
and wait if the TA is currently helping another student. When there are no students who
need help during office hours, the TA sits at the desk and takes a nap. If a student arrives
during office hours and finds the TA sleeping, the student must awaken the TA to ask for
help. If a student arrives and finds the TA currently helping another student, the student
sits on one of the chairs in the hallway and waits. If no chairs are available, the student
will come back at a later time.
Using threads, mutex locks, and semaphores, implement a solution that coordinates the
activities of the TA and the students.